# webhook/api/swagger-codegen/lettuce-webhook-openapi.yaml
openapi: 3.0.3
info:
  title: Plants Soil Lettuce (Webhook Engine)  - OpenAPI 3.0
  description: |-
    This is a Webhook Engine Core API based on the OpenAPI 3.0 specification.
    You can register as publisher and trigger (publish) events (message carried) as needed, all subscribers could receive the events instantly.
    You can register as subscriber with webhooks to listen on and receive events (messages carried) from publishers you subscribed.

    Some useful links:
    - [The Lettuce Webhook Engine repository](https://github.com/plants-soil/lettuce-webhook-hub)
  contact:
    email: danialdy@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://lettuce.plantssoil.com/api/v1
...
paths:
  /organization:
    put:
      tags:
        - organization
      summary: Update an existing organization
      description: Update an existing organization by organizationId
      operationId: updateOrganization
      requestBody:
        description: Update an existent organization
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
          application/xml:
            schema:
              $ref: '#/components/schemas/Organization'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Organization'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Required attributes cant't be null!
        '404':
          description: Organization not found
        '422':
          description: Validation exception
      security:
        - api_key: []
        - organization_auth:
            - write:organizations
            - read:organizations
    post:
      tags:
        - organization
      summary: Add a new organization
      description: Add a new organization
      operationId: addOrganization
      requestBody:
        description: Create a new organization
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
          application/xml:
            schema:
              $ref: '#/components/schemas/Organization'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Organization'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
      security:
        - api_key: []
        - organization_auth:
            - write:organizations
            - read:organizations
  /organization/all:
    get:
      tags:
        - organization
      summary: Find all organizations with pagination
      description: Will find the organizations on the page specified (page, pageSize)
      operationId: findAllOrganizations
      parameters:
        - name: page
          in: query
          description: The page number of organization query result
          required: true
          explode: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          description: Maximum organizations on current page
          required: true
          explode: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '400':
          description: Validation exception
      security:
        - organization_auth:
            - write:organizations
            - read:organizations
  /organization/{organizationId}:
    get:
      tags:
        - organization
      summary: Find organization by ID
      description: Returns a single organization
      operationId: findOrganizationById
      parameters:
        - name: organizationId
          in: path
          description: ID of organization to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid ID supplied
        '404':
          description: Organization not found
      security:
        - organization_auth:
            - write:organizations
            - read:organizations
    put:
      tags:
        - organization
      summary: Update an existing organization
      description: Update an existing organization by organizationId
      operationId: updateOrganizationById
      parameters:
        - name: organizationId
          in: path
          description: ID of organization that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent organization
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
          application/xml:
            schema:
              $ref: '#/components/schemas/Organization'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Organization'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Required attributes cant't be null!
        '404':
          description: Organization not found
        '422':
          description: Validation exception
      security:
        - api_key: []
        - organization_auth:
            - write:organizations
            - read:organizations
...
components:
  schemas:
    Organization:
      description: Could be company, business unit, or any other organization type, publisher / subscriber should belongs to one organization
      required:
        - organizationName
        - email
      type: object
      properties:
        organizationId:
          type: string
          description: ID of organization, should be unique, system will assign auto-generated ID if it's null
          example: 6791d42ed3d172552ed09459, maximum 24 chars
        organizationName:
          type: string
          description: Organization name, required
          example: Organiztion name
        email:
          type: string
          description: Organization official email, required and should be unique
          example: test@email.com, should be unique
        website:
          type: string
          description: Organization official website URL
          example: https://test.domain.com
        logoLink:
          type: string
          description: Organization official logo URL
          example: https://test.domain.com/images/logo.png
        organizationStatus:
          type: string
          description: Organization Status
          example: ACTIVE
          enum:
            - ACTIVE
            - INACTIVE
            - LOCKED
        createdBy:
          type: string
          description: The user name who create current organization
          example: danialdy
        creationTime:
          type: string
          description: The time when current organization is created
          format: date-time
          example: 2025-01-23T14:56:07Z
      xml:
        name: organization
...
  securitySchemes:
    organization_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:organizations: modify organizations in your account
            read:organizations: read your organizations
    publisher_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:publishers: modify publishers in you account
            read:publishers: read your publishers
    event_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:events: modify events in your account
            read:events: read your events
    datagroup_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:datagroups: modify data groups in your account
            read:datagroups: read your data groups
    subscriber_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:subscribers: modify subscribers in your account
            read:subscribers: read your subscribers
...